server/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=10.0.0,<11.0.0
python-multipart